requirements.txt
lxml
Pillow
requests
